move code to common file
