+++ src/libs/workflows-app-utils.js
@@ -1,0 +1,19 @@
+export const resolveRunningCromwellAppUrl = (apps, currentUser) => {
+  // it looks for Kubernetes deployment status RUNNING expressed by Leo
+  // See here for specific enumerations -- https://github.com/DataBiosphere/leonardo/blob/develop/core/src/main/scala/org/broadinstitute/dsde/workbench/leonardo/kubernetesModels.scala
+  // We explicitly look for a RUNNING app because if the CBAS app is not Running, we won't be able to send import method request.
+  const healthyState = 'RUNNING';
+  const cromwellAppType = 'CROMWELL';
+
+  // note: the requirement for checking if the app was created by user will not be needed when we move to multi-user Workflows app where users with
+  // OWNER and WRITER roles will be able to import methods to app created by another user
+  const filteredApps = apps.filter((app) => app.appType === cromwellAppType && app.status === healthyState && app.auditInfo.creator === currentUser);
+  if (filteredApps.length === 1) {
+    return {
+      cbasUrl: filteredApps[0].proxyUrls.cbas,
+      cbasUiUrl: filteredApps[0].proxyUrls['cbas-ui'],
+    };
+  }
+  // if there are no Running Cromwell apps or if there are more than one then it's an error state and return null
+  return null;
+};
